Rename generated (#433)
* generated code has 'generated' in the path

* regenerate code
