--- pkg/client/listers/servicecatalog/internalversion/expansion_generated.go
@@ -1,43 +1,0 @@
-/*
-Copyright 2017 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-// This file was automatically generated by lister-gen with arguments: --go-header-file=vendor/github.com/kubernetes/repo-infra/verify/boilerplate/boilerplate.go.txt --input-dirs=[github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog,github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1alpha1] --output-package=github.com/kubernetes-incubator/service-catalog/pkg/client/listers
-
-package internalversion
-
-// BindingListerExpansion allows custom methods to be added to
-// BindingLister.
-type BindingListerExpansion interface{}
-
-// BindingNamespaceListerExpansion allows custom methods to be added to
-// BindingNamespaeLister.
-type BindingNamespaceListerExpansion interface{}
-
-// BrokerListerExpansion allows custom methods to be added to
-// BrokerLister.
-type BrokerListerExpansion interface{}
-
-// InstanceListerExpansion allows custom methods to be added to
-// InstanceLister.
-type InstanceListerExpansion interface{}
-
-// InstanceNamespaceListerExpansion allows custom methods to be added to
-// InstanceNamespaeLister.
-type InstanceNamespaceListerExpansion interface{}
-
-// ServiceClassListerExpansion allows custom methods to be added to
-// ServiceClassLister.
-type ServiceClassListerExpansion interface{}
