+++ pkg/controller/wip/controller_test.go
@@ -24,8 +24,8 @@ import (
 	"github.com/kubernetes-incubator/service-catalog/pkg/brokerapi"
 	fakebrokerapi "github.com/kubernetes-incubator/service-catalog/pkg/brokerapi/fake"
 	servicecatalogclientset "github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/clientset/fake"
-	servicecataloginformers "github.com/kubernetes-incubator/service-catalog/pkg/client/informers"
-	v1alpha1informers "github.com/kubernetes-incubator/service-catalog/pkg/client/informers/servicecatalog/v1alpha1"
+	servicecataloginformers "github.com/kubernetes-incubator/service-catalog/pkg/client/informers_generated"
+	v1alpha1informers "github.com/kubernetes-incubator/service-catalog/pkg/client/informers_generated/servicecatalog/v1alpha1"
 
 	"k8s.io/client-go/1.5/kubernetes/fake"
